[GB.GTK] * NEW: The Application.Tooltip.Delay property was removed.
[GB.QT]
* NEW: The Application.Tooltip.Delay property was removed.

[GB.QT4]
* NEW: Backport features from gb.qt.
* BUG: The background and foreground color management is almost fixed.
* BUG: Panel frames is fixed.
* BUG: ToolButton icon size is fixed.
* BUG: ListBox background is fixed.
* BUG: GridView grid color is fixed.

[GB.QT4.EXT]
* NEW: Backport features from gb.qt.ext.
* NEW: Some source files are now shared with gb.qt.ext.


git-svn-id: svn://localhost/gambas/trunk@2009 867c0c6c-44f3-4631-809d-bfa615b0a4ec

+++ gb.qt/src/ext/CEditor.cpp
@@ -248,7 +248,7 @@ static void print_text(void *_object, const char *str, int lstr, bool esc = fals
 				len = MAX_CONSOLE_WIDTH - col;
 			WIDGET->insert(s.mid(i, len));
 			i += len;	
-			if (i >= s.length())
+			if (i >= (uint)s.length())
 				break;
 			col += len;
 		}
